Separate all rendering tasks into dedicated classes
Move the transform logic to the Transform handler.
Provide unit tests for the Zoom handler.
