+++ utiliti/src/de/gurkenlabs/utiliti/components/MapComponent.java
@@ -1,28 +1,19 @@
 package de.gurkenlabs.utiliti.components;
 
-import java.awt.BasicStroke;
 import java.awt.Color;
-import java.awt.Font;
-import java.awt.FontMetrics;
 import java.awt.Graphics2D;
-import java.awt.Shape;
-import java.awt.Stroke;
 import java.awt.event.FocusAdapter;
 import java.awt.event.FocusEvent;
 import java.awt.event.KeyEvent;
 import java.awt.event.MouseEvent;
-import java.awt.geom.Ellipse2D;
-import java.awt.geom.Path2D;
 import java.awt.geom.Point2D;
 import java.awt.geom.Rectangle2D;
-import java.awt.geom.RoundRectangle2D;
 import java.awt.image.BufferedImage;
 import java.nio.file.Paths;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
-import java.util.Map.Entry;
 import java.util.Optional;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.CopyOnWriteArrayList;
@@ -34,11 +25,7 @@ import java.util.logging.Logger;
 import javax.swing.JOptionPane;
 import javax.swing.SwingUtilities;
 
-import de.gurkenlabs.litiengine.Align;
 import de.gurkenlabs.litiengine.Game;
-import de.gurkenlabs.litiengine.IUpdateable;
-import de.gurkenlabs.litiengine.Valign;
-import de.gurkenlabs.litiengine.entities.CollisionEntity;
 import de.gurkenlabs.litiengine.entities.IEntity;
 import de.gurkenlabs.litiengine.environment.Environment;
 import de.gurkenlabs.litiengine.environment.tilemap.IImageLayer;
@@ -48,16 +35,13 @@ import de.gurkenlabs.litiengine.environment.tilemap.IMapObjectLayer;
 import de.gurkenlabs.litiengine.environment.tilemap.ITileset;
 import de.gurkenlabs.litiengine.environment.tilemap.MapObjectProperty;
 import de.gurkenlabs.litiengine.environment.tilemap.MapObjectType;
-import de.gurkenlabs.litiengine.environment.tilemap.MapUtilities;
 import de.gurkenlabs.litiengine.environment.tilemap.xml.Blueprint;
 import de.gurkenlabs.litiengine.environment.tilemap.xml.MapObject;
 import de.gurkenlabs.litiengine.environment.tilemap.xml.MapObjectLayer;
 import de.gurkenlabs.litiengine.environment.tilemap.xml.Tileset;
 import de.gurkenlabs.litiengine.environment.tilemap.xml.TmxMap;
 import de.gurkenlabs.litiengine.graphics.ImageFormat;
-import de.gurkenlabs.litiengine.graphics.RenderEngine;
 import de.gurkenlabs.litiengine.graphics.Spritesheet;
-import de.gurkenlabs.litiengine.graphics.TextRenderer;
 import de.gurkenlabs.litiengine.gui.ComponentMouseEvent;
 import de.gurkenlabs.litiengine.gui.ComponentMouseWheelEvent;
 import de.gurkenlabs.litiengine.gui.GuiComponent;
@@ -65,24 +49,23 @@ import de.gurkenlabs.litiengine.input.Input;
 import de.gurkenlabs.litiengine.physics.Collision;
 import de.gurkenlabs.litiengine.resources.Resources;
 import de.gurkenlabs.litiengine.resources.SpritesheetResource;
-import de.gurkenlabs.litiengine.util.ColorHelper;
 import de.gurkenlabs.litiengine.util.MathUtilities;
 import de.gurkenlabs.litiengine.util.geom.GeometricUtilities;
 import de.gurkenlabs.litiengine.util.io.FileUtilities;
 import de.gurkenlabs.litiengine.util.io.ImageSerializer;
 import de.gurkenlabs.utiliti.Cursors;
 import de.gurkenlabs.utiliti.Program;
-import de.gurkenlabs.utiliti.Style;
 import de.gurkenlabs.utiliti.UndoManager;
 import de.gurkenlabs.utiliti.handlers.Snap;
-import de.gurkenlabs.utiliti.handlers.Zoom;
-import de.gurkenlabs.utiliti.handlers.Transform.ResizeAnchor;
+import de.gurkenlabs.utiliti.handlers.Transform;
 import de.gurkenlabs.utiliti.handlers.Transform.TransformType;
+import de.gurkenlabs.utiliti.handlers.Zoom;
+import de.gurkenlabs.utiliti.renderers.Renderers;
 import de.gurkenlabs.utiliti.swing.UI;
 import de.gurkenlabs.utiliti.swing.dialogs.XmlExportDialog;
 import de.gurkenlabs.utiliti.swing.dialogs.XmlImportDialog;
 
-public class MapComponent extends GuiComponent implements IUpdateable {
+public class MapComponent extends GuiComponent {
 
   public static final int EDITMODE_CREATE = 0;
   public static final int EDITMODE_EDIT = 1;
@@ -95,11 +78,8 @@ public class MapComponent extends GuiComponent implements IUpdateable {
   private static final Logger log = Logger.getLogger(MapComponent.class.getName());
 
   private static final String DEFAULT_MAPOBJECTLAYER_NAME = "default";
-  private static final int TRANSFORM_RECT_SIZE = 6;
-  private static final int BASE_SCROLL_SPEED = 50;
 
-  private double currentTransformRectSize = TRANSFORM_RECT_SIZE;
-  private final java.util.Map<ResizeAnchor, Rectangle2D> transformRects;
+  private static final int BASE_SCROLL_SPEED = 50;
 
   private final List<IntConsumer> editModeChangedConsumer;
   private final List<Consumer<IMapObject>> focusChangedConsumer;
@@ -114,9 +94,7 @@ public class MapComponent extends GuiComponent implements IUpdateable {
   private final java.util.Map<String, Environment> environments;
   private final java.util.Map<IMapObject, Point2D> dragLocationMapObjects;
 
-  private int currentEditMode = EDITMODE_EDIT;
-  private TransformType currentTransform;
-  private ResizeAnchor currentAnchor;
+  private int editMode = EDITMODE_EDIT;
 
   private final List<TmxMap> maps;
 
@@ -131,12 +109,7 @@ public class MapComponent extends GuiComponent implements IUpdateable {
   private boolean isFocussing;
   private float dragSizeHeight;
   private float dragSizeWidth;
-  private Rectangle2D newObjectArea;
   private Blueprint copiedBlueprint;
-
-  private Color colorSelectionBorder;
-  private float focusBorderBrightness = 0;
-  private boolean focusBorderBrightnessIncreasing = true;
 
   private final EditorScreen screen;
 
@@ -156,11 +129,10 @@ public class MapComponent extends GuiComponent implements IUpdateable {
     this.environments = new ConcurrentHashMap<>();
     this.maps = new ArrayList<>();
     this.cameraFocus = new ConcurrentHashMap<>();
-    this.transformRects = new ConcurrentHashMap<>();
     this.dragLocationMapObjects = new ConcurrentHashMap<>();
     this.screen = screen;
 
-    UndoManager.onUndoStackChanged(e -> this.updateTransformControls());
+    UndoManager.onUndoStackChanged(e -> Transform.updateAnchors());
   }
 
   public void onEditModeChanged(IntConsumer cons) {
@@ -193,27 +165,7 @@ public class MapComponent extends GuiComponent implements IUpdateable {
       return;
     }
 
-    renderGrid(g);
-
-    final BasicStroke shapeStroke = new BasicStroke(1 / Game.world().camera().getRenderScale());
-    if (Program.preferences().isRenderBoundingBoxes()) {
-      renderMapObjectBounds(g);
-    }
-
-    switch (this.currentEditMode) {
-    case EDITMODE_CREATE:
-      this.renderNewObjectArea(g, shapeStroke);
-      break;
-    case EDITMODE_EDIT:
-      this.renderMouseSelectionArea(g, shapeStroke);
-      break;
-    default:
-      break;
-    }
-
-    this.renderSelection(g);
-    this.renderFocus(g);
-
+    Renderers.render(g);
     super.render(g);
   }
 
@@ -259,6 +211,15 @@ public class MapComponent extends GuiComponent implements IUpdateable {
     return null;
   }
 
+  public Rectangle2D getFocusBounds() {
+    final IMapObject focusedObject = this.getFocusedMapObject();
+    if (focusedObject == null) {
+      return null;
+    }
+
+    return focusedObject.getBoundingBox();
+  }
+
   public List<IMapObject> getSelectedMapObjects() {
     if (Game.world().environment() != null && Game.world().environment().getMap() != null) {
       final String map = Game.world().environment().getMap().getName();
@@ -281,8 +242,7 @@ public class MapComponent extends GuiComponent implements IUpdateable {
   @Override
   public void prepare() {
     Game.world().camera().onZoomChanged(zoom -> {
-      this.currentTransformRectSize = TRANSFORM_RECT_SIZE / zoom;
-      this.updateTransformControls();
+      Transform.updateAnchors();
       this.scrollSpeed = BASE_SCROLL_SPEED / zoom;
     });
 
@@ -298,8 +258,6 @@ public class MapComponent extends GuiComponent implements IUpdateable {
       this.setupKeyboardControls();
       this.setupMouseControls();
       this.initialized = true;
-
-      Game.loop().attach(this);
     }
 
     super.prepare();
@@ -380,7 +338,7 @@ public class MapComponent extends GuiComponent implements IUpdateable {
       this.setSelection(mapObject, false);
       this.setFocus(mapObject, false);
     }
-    this.updateTransformControls();
+    Transform.updateAnchors();
     this.setEditMode(EDITMODE_MOVE);
   }
 
@@ -553,7 +511,7 @@ public class MapComponent extends GuiComponent implements IUpdateable {
 
   public void centerCameraOnFocus() {
     if (this.hasFocus() && this.getFocusedMapObject() != null) {
-      final Rectangle2D focus = this.getFocus();
+      final Rectangle2D focus = this.getFocusBounds();
       if (focus == null) {
         return;
       }
@@ -571,8 +529,12 @@ public class MapComponent extends GuiComponent implements IUpdateable {
     Game.world().camera().setFocus(env.getCenter());
   }
 
+  public int getEditMode() {
+    return this.editMode;
+  }
+
   public void setEditMode(int editMode) {
-    if (editMode == this.currentEditMode) {
+    if (editMode == this.editMode) {
       return;
     }
 
@@ -592,9 +554,9 @@ public class MapComponent extends GuiComponent implements IUpdateable {
       break;
     }
 
-    this.currentEditMode = editMode;
+    this.editMode = editMode;
     for (IntConsumer cons : this.editModeChangedConsumer) {
-      cons.accept(this.currentEditMode);
+      cons.accept(this.editMode);
     }
   }
 
@@ -630,7 +592,7 @@ public class MapComponent extends GuiComponent implements IUpdateable {
         cons.accept(mapObject);
       }
 
-      this.updateTransformControls();
+      Transform.updateAnchors();
       this.setSelection(mapObject, clearSelection);
     } finally {
       this.isFocussing = false;
@@ -667,19 +629,6 @@ public class MapComponent extends GuiComponent implements IUpdateable {
 
     for (Consumer<List<IMapObject>> cons : this.selectionChangedConsumer) {
       cons.accept(this.getSelectedMapObjects());
-    }
-  }
-
-  public void updateTransformControls() {
-    final Rectangle2D focus = this.getFocus();
-    if (focus == null) {
-      this.transformRects.clear();
-      return;
-    }
-
-    for (ResizeAnchor trans : ResizeAnchor.values()) {
-      Rectangle2D transRect = new Rectangle2D.Double(this.getTransX(trans, focus), this.getTransY(trans, focus), this.currentTransformRectSize, this.currentTransformRectSize);
-      this.transformRects.put(trans, transRect);
     }
   }
 
@@ -844,23 +793,6 @@ public class MapComponent extends GuiComponent implements IUpdateable {
   }
 
   @Override
-  public void update() {
-    if (this.focusBorderBrightness <= 0.4) {
-      this.focusBorderBrightnessIncreasing = true;
-    } else if (this.focusBorderBrightness >= 0.9) {
-      this.focusBorderBrightnessIncreasing = false;
-    }
-
-    if (this.focusBorderBrightnessIncreasing && this.focusBorderBrightness < 0.9) {
-      this.focusBorderBrightness += 0.01;
-    } else if (!focusBorderBrightnessIncreasing && this.focusBorderBrightness >= 0.4) {
-      this.focusBorderBrightness -= 0.01;
-    }
-
-    this.colorSelectionBorder = Color.getHSBColor(0, 0, this.focusBorderBrightness);
-  }
-
-  @Override
   protected boolean mouseEventShouldBeForwarded(final MouseEvent e) {
     return this.isForwardMouseEvents() && this.isVisible() && this.isEnabled() && !this.isSuspended() && e != null;
   }
@@ -870,14 +802,15 @@ public class MapComponent extends GuiComponent implements IUpdateable {
   }
 
   private IMapObject createNewMapObject(MapObjectType type) {
+    final Rectangle2D newObjectArea = this.getMouseSelectArea(true);
     IMapObject mo = new MapObject();
     mo.setType(type.toString());
-    mo.setX((float) this.newObjectArea.getX());
-    mo.setY((float) this.newObjectArea.getY());
+    mo.setX((float) newObjectArea.getX());
+    mo.setY((float) newObjectArea.getY());
 
     // ensure a minimum size for the new object
-    float width = (float) this.newObjectArea.getWidth();
-    float height = (float) this.newObjectArea.getHeight();
+    float width = (float) newObjectArea.getWidth();
+    float height = (float) newObjectArea.getHeight();
     mo.setWidth(width == 0 ? 16 : width);
     mo.setHeight(height == 0 ? 16 : height);
     mo.setId(Game.world().environment().getNextMapId());
@@ -885,8 +818,8 @@ public class MapComponent extends GuiComponent implements IUpdateable {
 
     switch (type) {
     case PROP:
-      mo.setValue(MapObjectProperty.COLLISIONBOX_WIDTH, (this.newObjectArea.getWidth() * 0.4));
-      mo.setValue(MapObjectProperty.COLLISIONBOX_HEIGHT, (this.newObjectArea.getHeight() * 0.4));
+      mo.setValue(MapObjectProperty.COLLISIONBOX_WIDTH, (newObjectArea.getWidth() * 0.4));
+      mo.setValue(MapObjectProperty.COLLISIONBOX_HEIGHT, (newObjectArea.getHeight() * 0.4));
       mo.setValue(MapObjectProperty.COLLISION, true);
       mo.setValue(MapObjectProperty.COMBAT_INDESTRUCTIBLE, false);
       mo.setValue(MapObjectProperty.PROP_ADDSHADOW, true);
@@ -908,7 +841,7 @@ public class MapComponent extends GuiComponent implements IUpdateable {
     return mo;
   }
 
-  private Rectangle2D getCurrentMouseSelectionArea(boolean snap) {
+  public Rectangle2D getMouseSelectArea(boolean snap) {
     final Point2D start = this.startPoint;
     if (start == null) {
       return null;
@@ -941,54 +874,9 @@ public class MapComponent extends GuiComponent implements IUpdateable {
     return new Rectangle2D.Double(minX, minY, width, height);
   }
 
-  private Rectangle2D getFocus() {
-    final IMapObject focusedObject = this.getFocusedMapObject();
-    if (focusedObject == null) {
-      return null;
-    }
-
-    return focusedObject.getBoundingBox();
-  }
-
-  private double getTransX(ResizeAnchor type, Rectangle2D focus) {
-    switch (type) {
-    case DOWN:
-    case UP:
-      return focus.getCenterX() - this.currentTransformRectSize / 2;
-    case LEFT:
-    case DOWNLEFT:
-    case UPLEFT:
-      return focus.getX() - this.currentTransformRectSize;
-    case RIGHT:
-    case DOWNRIGHT:
-    case UPRIGHT:
-      return focus.getMaxX();
-    default:
-      return 0;
-    }
-  }
-
-  private double getTransY(ResizeAnchor type, Rectangle2D focus) {
-    switch (type) {
-    case DOWN:
-    case DOWNLEFT:
-    case DOWNRIGHT:
-      return focus.getMaxY();
-    case UP:
-    case UPLEFT:
-    case UPRIGHT:
-      return focus.getY() - this.currentTransformRectSize;
-    case LEFT:
-    case RIGHT:
-      return focus.getCenterY() - this.currentTransformRectSize / 2;
-    default:
-      return 0;
-    }
-  }
-
-  private void handleTransform() {
+  private void handleResizeTransform() {
     final IMapObject transformObject = this.getFocusedMapObject();
-    if (transformObject == null || this.currentEditMode != EDITMODE_EDIT || currentTransform != TransformType.RESIZE || this.currentAnchor == null) {
+    if (transformObject == null || this.editMode != EDITMODE_EDIT || Transform.type() != TransformType.RESIZE || Transform.anchor() == null) {
       return;
     }
 
@@ -1008,7 +896,7 @@ public class MapComponent extends GuiComponent implements IUpdateable {
     double newX = dragLocationMapObject.getX();
     double newY = dragLocationMapObject.getY();
 
-    switch (this.currentAnchor) {
+    switch (Transform.anchor()) {
     case DOWN:
       newHeight += deltaY;
       break;
@@ -1084,10 +972,10 @@ public class MapComponent extends GuiComponent implements IUpdateable {
     }
 
     UI.getInspector().bind(transformObject);
-    this.updateTransformControls();
+    Transform.updateAnchors();
   }
 
-  private void handleSelectedEntitiesDrag() {
+  private void handleMoveTransform() {
     if (!this.isMoving) {
       this.isMoving = true;
 
@@ -1109,7 +997,7 @@ public class MapComponent extends GuiComponent implements IUpdateable {
       }
     }
 
-    if (minX == null || minY == null || (!Input.keyboard().isPressed(KeyEvent.VK_CONTROL) && this.currentEditMode != EDITMODE_MOVE)) {
+    if (minX == null || minY == null || (!Input.keyboard().isPressed(KeyEvent.VK_CONTROL) && this.editMode != EDITMODE_MOVE)) {
       return;
     }
 
@@ -1181,7 +1069,7 @@ public class MapComponent extends GuiComponent implements IUpdateable {
 
       if (selected.equals(this.getFocusedMapObject())) {
         UI.getInspector().bind(selected);
-        this.updateTransformControls();
+        Transform.updateAnchors();
       }
     }
   }
@@ -1195,19 +1083,19 @@ public class MapComponent extends GuiComponent implements IUpdateable {
 
   private void setupKeyboardControls() {
     Input.keyboard().onKeyPressed(KeyEvent.VK_CONTROL, e -> {
-      if (this.currentEditMode == EDITMODE_EDIT && this.getFocus() != null) {
+      if (this.editMode == EDITMODE_EDIT && this.getFocusBounds() != null) {
         this.setEditMode(EDITMODE_MOVE);
       }
     });
 
     Input.keyboard().onKeyReleased(KeyEvent.VK_CONTROL, e -> {
-      if (this.currentEditMode == EDITMODE_MOVE) {
+      if (this.editMode == EDITMODE_MOVE) {
         this.setEditMode(EDITMODE_EDIT);
       }
     });
 
     Input.keyboard().onKeyPressed(KeyEvent.VK_ESCAPE, e -> {
-      if (this.currentEditMode == EDITMODE_CREATE) {
+      if (this.editMode == EDITMODE_CREATE) {
         this.setEditMode(EDITMODE_EDIT);
       }
     });
@@ -1285,7 +1173,7 @@ public class MapComponent extends GuiComponent implements IUpdateable {
   }
 
   private static void afterKeyPressed() {
-    EditorScreen.instance().getMapComponent().updateTransformControls();
+    Transform.updateAnchors();
   }
 
   private void setupMouseControls() {
@@ -1350,50 +1238,7 @@ public class MapComponent extends GuiComponent implements IUpdateable {
    *          The mouse event of the calling {@link GuiComponent}
    */
   private void handleMouseMoved(ComponentMouseEvent e) {
-    boolean hovered = false;
-    if (this.getFocus() != null) {
-      this.currentAnchor = null;
-
-      for (Entry<ResizeAnchor, Rectangle2D> entry : this.transformRects.entrySet()) {
-        Rectangle2D rect = entry.getValue();
-        Rectangle2D hoverrect = new Rectangle2D.Double(rect.getX() - rect.getWidth() * 2.5, rect.getY() - rect.getHeight() * 2.5, rect.getWidth() * 5, rect.getHeight() * 5);
-        if (hoverrect.contains(Input.mouse().getMapLocation())) {
-          hovered = true;
-          if (entry.getKey() == ResizeAnchor.DOWN || entry.getKey() == ResizeAnchor.UP) {
-            Game.window().getRenderComponent().setCursor(Cursors.TRANS_VERTICAL, 0, 0);
-          } else if (entry.getKey() == ResizeAnchor.UPLEFT || entry.getKey() == ResizeAnchor.DOWNRIGHT) {
-            Game.window().getRenderComponent().setCursor(Cursors.TRANS_DIAGONAL_LEFT, 0, 0);
-          } else if (entry.getKey() == ResizeAnchor.UPRIGHT || entry.getKey() == ResizeAnchor.DOWNLEFT) {
-            Game.window().getRenderComponent().setCursor(Cursors.TRANS_DIAGONAL_RIGHT, 0, 0);
-          } else {
-            Game.window().getRenderComponent().setCursor(Cursors.TRANS_HORIZONTAL, 0, 0);
-          }
-
-          this.currentAnchor = entry.getKey();
-          this.currentTransform = TransformType.RESIZE;
-          break;
-        }
-      }
-    }
-
-    if (!hovered) {
-      boolean moveMode = false;
-
-      for (IMapObject selected : this.getSelectedMapObjects()) {
-        if (selected.getBoundingBox().contains(Input.mouse().getMapLocation())) {
-          moveMode = true;
-          break;
-        }
-      }
-
-      if (moveMode) {
-        Game.window().getRenderComponent().setCursor(Cursors.MOVE, 0, 0);
-        this.currentTransform = TransformType.MOVE;
-      } else {
-        Game.window().getRenderComponent().setCursor(Cursors.DEFAULT, 0, 0);
-        this.currentTransform = TransformType.NONE;
-      }
-    }
+    Transform.updateTransform();
   }
 
   private void handleMousePressed(ComponentMouseEvent e) {
@@ -1401,11 +1246,11 @@ public class MapComponent extends GuiComponent implements IUpdateable {
       return;
     }
 
-    if (this.currentTransform == TransformType.MOVE) {
+    if (Transform.type() == TransformType.MOVE) {
       this.setEditMode(EDITMODE_MOVE);
     }
 
-    switch (this.currentEditMode) {
+    switch (this.editMode) {
     case EDITMODE_CREATE:
     case EDITMODE_MOVE:
       if (SwingUtilities.isLeftMouseButton(e.getEvent())) {
@@ -1413,7 +1258,7 @@ public class MapComponent extends GuiComponent implements IUpdateable {
       }
       break;
     case EDITMODE_EDIT:
-      if (this.isMoving || this.currentTransform == TransformType.RESIZE || SwingUtilities.isRightMouseButton(e.getEvent())) {
+      if (this.isMoving || Transform.type() == TransformType.RESIZE || SwingUtilities.isRightMouseButton(e.getEvent())) {
         return;
       }
 
@@ -1430,29 +1275,19 @@ public class MapComponent extends GuiComponent implements IUpdateable {
       return;
     }
 
-    switch (this.currentEditMode) {
-    case EDITMODE_CREATE:
-      if (startPoint == null) {
-        return;
-      }
-
-      if (SwingUtilities.isLeftMouseButton(e.getEvent())) {
-        newObjectArea = this.getCurrentMouseSelectionArea(true);
-      }
-
-      break;
+    switch (this.editMode) {
     case EDITMODE_EDIT:
-      if (this.currentTransform == TransformType.RESIZE) {
+      if (Transform.type() == TransformType.RESIZE) {
         if (!this.isTransforming) {
           this.isTransforming = true;
           UndoManager.instance().mapObjectChanging(this.getFocusedMapObject());
         }
 
-        this.handleTransform();
+        this.handleResizeTransform();
       }
       break;
     case EDITMODE_MOVE:
-      this.handleSelectedEntitiesDrag();
+      this.handleMoveTransform();
 
       break;
     default:
@@ -1470,20 +1305,15 @@ public class MapComponent extends GuiComponent implements IUpdateable {
     this.dragSizeHeight = 0;
     this.dragSizeWidth = 0;
 
-    switch (this.currentEditMode) {
+    switch (this.editMode) {
     case EDITMODE_CREATE:
       if (SwingUtilities.isRightMouseButton(e.getEvent())) {
-        this.newObjectArea = null;
         this.setEditMode(EDITMODE_EDIT);
         break;
       }
 
-      if (this.newObjectArea == null) {
-        break;
-      }
-
       IMapObject mo = this.createNewMapObject(UI.getInspector().getObjectType());
-      this.newObjectArea = null;
+
       this.setFocus(mo, !Input.keyboard().isPressed(KeyEvent.VK_SHIFT));
       UI.getInspector().bind(mo);
       this.setEditMode(EDITMODE_EDIT);
@@ -1526,7 +1356,7 @@ public class MapComponent extends GuiComponent implements IUpdateable {
   }
 
   private void evaluateFocus() {
-    Rectangle2D rect = this.getCurrentMouseSelectionArea(false);
+    Rectangle2D rect = this.getMouseSelectArea(false);
     if (rect == null) {
       return;
     }
@@ -1590,333 +1420,5 @@ public class MapComponent extends GuiComponent implements IUpdateable {
     }
 
     return true;
-  }
-
-  private static void renderMapObjectBounds(Graphics2D g) {
-    if (Game.world().environment() == null || Game.world().environment().getMap() == null) {
-      return;
-    }
-
-    final List<IMapObjectLayer> layers = Game.world().environment().getMap().getMapObjectLayers();
-    // render all entities
-    for (final IMapObjectLayer layer : layers) {
-      if (layer == null || !layer.isVisible()) {
-        continue;
-      }
-
-      for (final IMapObject mapObject : layer.getMapObjects()) {
-        if (mapObject == null) {
-          continue;
-        }
-
-        MapObjectType type = MapObjectType.get(mapObject.getType());
-        final BasicStroke shapeStroke = new BasicStroke(1f / Game.world().camera().getRenderScale());
-        if (type == null) {
-          if (Program.preferences().isRenderCustomMapObjects()) {
-            renderUnsupportedMapObject(g, mapObject, shapeStroke);
-          }
-
-          continue;
-        }
-
-        // render spawn points
-        if (type == MapObjectType.SPAWNPOINT) {
-          g.setColor(Style.COLOR_SPAWNPOINT);
-          RenderEngine.renderShape(g, new Rectangle2D.Double(mapObject.getBoundingBox().getCenterX() - 1, mapObject.getBoundingBox().getCenterY() - 1, 2, 2));
-        }
-
-        if (type != MapObjectType.COLLISIONBOX) {
-          Color colorBoundingBoxFill;
-          if (layer.getColor() != null) {
-            colorBoundingBoxFill = new Color(layer.getColor().getRed(), layer.getColor().getGreen(), layer.getColor().getBlue(), 15);
-          } else {
-            colorBoundingBoxFill = Style.COLOR_DEFAULT_BOUNDING_BOX_FILL;
-          }
-
-          renderBoundingBox(g, mapObject, colorBoundingBoxFill, shapeStroke);
-        }
-
-        renderCollisionBox(g, mapObject, shapeStroke);
-      }
-    }
-  }
-
-  private static void renderUnsupportedMapObject(Graphics2D g, IMapObject mapObject, BasicStroke shapeStroke) {
-    g.setColor(Style.COLOR_UNSUPPORTED);
-    Point2D start = new Point2D.Double(mapObject.getLocation().getX(), mapObject.getLocation().getY());
-    StringBuilder info = new StringBuilder("#");
-    info.append(mapObject.getId());
-    if (mapObject.getName() != null && !mapObject.getName().isEmpty()) {
-      info.append("(");
-      info.append(mapObject.getName());
-      info.append(")");
-    }
-
-    RenderEngine.renderText(g, info.toString(), start.getX(), start.getY() - 5);
-    RenderEngine.renderShape(g, new Ellipse2D.Double(start.getX() - 1, start.getY() - 1, 3, 3));
-
-    if (mapObject.isPolyline()) {
-
-      if (mapObject.getPolyline() == null || mapObject.getPolyline().getPoints().isEmpty()) {
-        return;
-      }
-
-      // found the path for the rat
-      final Path2D path = MapUtilities.convertPolyshapeToPath(mapObject);
-      if (path == null) {
-        return;
-      }
-
-      RenderEngine.renderOutline(g, path, shapeStroke);
-    } else if (mapObject.isPolygon()) {
-      if (mapObject.getPolygon() == null || mapObject.getPolygon().getPoints().isEmpty()) {
-        return;
-      }
-
-      // found the path for the rat
-      final Path2D path = MapUtilities.convertPolyshapeToPath(mapObject);
-      if (path == null) {
-        return;
-      }
-
-      g.setColor(Style.COLOR_UNSUPPORTED_FILL);
-      RenderEngine.renderShape(g, path);
-      g.setColor(Style.COLOR_UNSUPPORTED);
-      RenderEngine.renderOutline(g, path, shapeStroke);
-    } else if (mapObject.isEllipse()) {
-      if (mapObject.getEllipse() == null) {
-        return;
-      }
-      g.setColor(Style.COLOR_UNSUPPORTED_FILL);
-      RenderEngine.renderShape(g, mapObject.getEllipse());
-
-      g.setColor(Style.COLOR_UNSUPPORTED);
-      RenderEngine.renderOutline(g, mapObject.getEllipse(), shapeStroke);
-    } else {
-      g.setColor(Style.COLOR_UNSUPPORTED_FILL);
-      RenderEngine.renderShape(g, mapObject.getBoundingBox());
-      g.setColor(Style.COLOR_UNSUPPORTED);
-      RenderEngine.renderOutline(g, mapObject.getBoundingBox(), shapeStroke);
-    }
-  }
-
-  private static void renderName(Graphics2D g, IMapObject mapObject) {
-    g.setFont(Style.FONT_DEFAULT.deriveFont(10f));
-    FontMetrics fm = g.getFontMetrics();
-
-    String objectName = mapObject.getName();
-    if (objectName != null && !objectName.isEmpty()) {
-      final int PADDING = 2;
-      double stringWidth = fm.stringWidth(objectName) / Game.world().camera().getRenderScale();
-      double stringHeight = fm.getHeight() * .5 / Game.world().camera().getRenderScale();
-      double x = mapObject.getX() + ((mapObject.getWidth() - stringWidth) / 2.0) - PADDING;
-      double y = mapObject.getY() + mapObject.getHeight() + stringHeight;
-      double width = stringWidth + PADDING * 2;
-      double height = stringHeight + PADDING * 2;
-      RoundRectangle2D rect = new RoundRectangle2D.Double(x, y, width, height, 2, 2);
-      g.setColor(new Color(0, 0, 0, 200));
-      RenderEngine.renderShape(g, rect, true);
-
-      g.setColor(Color.WHITE);
-      RenderEngine.renderText(g, mapObject.getName(), x + PADDING, rect.getMaxY() - PADDING, true);
-    }
-  }
-
-  private static void renderGrid(Graphics2D g) {
-    // render the grid
-    if (Program.preferences().isShowGrid() && Game.world().camera().getRenderScale() >= 1 && Game.world().environment() != null) {
-
-      final IMap map = Game.world().environment().getMap();
-      if (map == null) {
-        return;
-      }
-
-      g.setColor(Program.preferences().getGridColor());
-      final Stroke stroke = new BasicStroke(Program.preferences().getGridLineWidth() / Game.world().camera().getRenderScale());
-      for (int x = 0; x < map.getWidth(); x++) {
-        for (int y = 0; y < map.getHeight(); y++) {
-          Shape tile = map.getOrientation().getShape(x, y, map);
-          if (Game.world().camera().getViewport().intersects(tile.getBounds2D())) {
-            RenderEngine.renderOutline(g, tile, stroke);
-          }
-        }
-      }
-    }
-  }
-
-  private void renderNewObjectArea(Graphics2D g, Stroke shapeStroke) {
-    if (this.newObjectArea == null) {
-      return;
-    }
-
-    g.setColor(Style.COLOR_NEWOBJECT_FILL);
-    RenderEngine.renderShape(g, newObjectArea);
-    g.setColor(Style.COLOR_NEWOBJECT_BORDER);
-    RenderEngine.renderOutline(g, newObjectArea, shapeStroke);
-    g.setFont(g.getFont().deriveFont(Font.BOLD));
-    RenderEngine.renderText(g, newObjectArea.getWidth() + "", newObjectArea.getX() + newObjectArea.getWidth() / 2 - 3, newObjectArea.getY() - 5);
-    RenderEngine.renderText(g, newObjectArea.getHeight() + "", newObjectArea.getX() - 10, newObjectArea.getY() + newObjectArea.getHeight() / 2);
-  }
-
-  private void renderMouseSelectionArea(Graphics2D g, Stroke shapeStroke) {
-    // draw mouse selection area
-    final Point2D start = this.startPoint;
-    if (start != null && !Input.keyboard().isPressed(KeyEvent.VK_CONTROL)) {
-      final Rectangle2D rect = this.getCurrentMouseSelectionArea(false);
-      if (rect == null) {
-        return;
-      }
-
-      g.setColor(Style.COLOR_MOUSE_SELECTION_AREA_FILL);
-      RenderEngine.renderShape(g, rect);
-      g.setColor(Style.COLOR_MOUSE_SELECTION_AREA_BORDER);
-      RenderEngine.renderOutline(g, rect, shapeStroke);
-    }
-  }
-
-  private void renderFocus(Graphics2D g) {
-    // render the focus and the transform rects
-    final Rectangle2D focus = this.getFocus();
-    final IMapObject focusedMapObject = this.getFocusedMapObject();
-    if (focus != null && focusedMapObject != null) {
-      Stroke stroke = new BasicStroke(1 / Game.world().camera().getRenderScale(), BasicStroke.CAP_ROUND, BasicStroke.JOIN_MITER, 4, new float[] { 1f, 1f }, Game.time().now() / 15);
-
-      g.setColor(Color.BLACK);
-
-      RenderEngine.renderOutline(g, focus, stroke);
-
-      Stroke whiteStroke = new BasicStroke(1 / Game.world().camera().getRenderScale(), BasicStroke.CAP_ROUND, BasicStroke.JOIN_MITER, 4, new float[] { 1f, 1f }, Game.time().now() / 15.0f - 1f);
-      g.setColor(Color.WHITE);
-      RenderEngine.renderOutline(g, focus, whiteStroke);
-
-      // render transform rects
-      if (this.currentEditMode != EDITMODE_MOVE) {
-        Stroke transStroke = new BasicStroke(1 / Game.world().camera().getRenderScale());
-        for (Rectangle2D trans : this.transformRects.values()) {
-          g.setColor(Style.COLOR_TRANSFORM_RECT_FILL);
-          RenderEngine.renderShape(g, trans);
-          g.setColor(Color.BLACK);
-          RenderEngine.renderOutline(g, trans, transStroke);
-        }
-      }
-    }
-
-    if (focusedMapObject != null) {
-      renderObjectId(g, focusedMapObject);
-    }
-  }
-
-  private void renderObjectId(Graphics2D g, IMapObject mapObject) {
-    if (!Program.preferences().isRenderMapIds()) {
-      return;
-    }
-
-    Font previousFont = g.getFont();
-    Font idFont = previousFont.deriveFont(Math.max(8f, (float) (10 * Math.sqrt(Game.world().camera().getRenderScale()))));
-    if (Zoom.get() > 1) {
-      idFont = idFont.deriveFont(Font.BOLD);
-    }
-
-    Point2D loc = Game.world().camera().getViewportLocation(new Point2D.Double(mapObject.getX() + mapObject.getWidth() / 2, mapObject.getY()));
-    g.setColor(Style.COLOR_STATUS);
-
-    g.setFont(idFont);
-    String id = Integer.toString(mapObject.getId());
-
-    double x = loc.getX() * Game.world().camera().getRenderScale() - g.getFontMetrics().stringWidth(id) / 2.0;
-    double y = loc.getY() * Game.world().camera().getRenderScale() - (g.getFontMetrics().getHeight() * .30);
-
-    if (Zoom.get() < 1) {
-      TextRenderer.render(g, id, x, y);
-    } else {
-      TextRenderer.renderWithOutline(g, id, x, y, Style.COLOR_DARKBORDER, 5, true);
-    }
-
-    g.setFont(previousFont);
-  }
-
-  private void renderSelection(Graphics2D g) {
-    for (IMapObject mapObject : this.getSelectedMapObjects()) {
-      if (mapObject.equals(this.getFocusedMapObject())) {
-        continue;
-      }
-
-      Stroke stroke = new BasicStroke(1 / Game.world().camera().getRenderScale());
-
-      g.setColor(colorSelectionBorder);
-      RenderEngine.renderOutline(g, mapObject.getBoundingBox(), stroke);
-      renderObjectId(g, mapObject);
-    }
-  }
-
-  private static void renderBoundingBox(Graphics2D g, IMapObject mapObject, Color colorBoundingBoxFill, BasicStroke shapeStroke) {
-    MapObjectType type = MapObjectType.get(mapObject.getType());
-    Color fillColor = colorBoundingBoxFill;
-    if (type == MapObjectType.TRIGGER) {
-      fillColor = Style.COLOR_TRIGGER_FILL;
-    } else if (type == MapObjectType.STATICSHADOW) {
-      fillColor = Style.COLOR_SHADOW_FILL;
-    }
-
-    // render bounding boxes
-    g.setColor(fillColor);
-
-    // don't fill rect for lightsource because it is important to judge
-    // the color
-    if (type != MapObjectType.LIGHTSOURCE) {
-      RenderEngine.renderShape(g, mapObject.getBoundingBox());
-    }
-
-    Color borderColor = colorBoundingBoxFill;
-    if (type == MapObjectType.TRIGGER) {
-      borderColor = Style.COLOR_TRIGGER_BORDER;
-    } else if (type == MapObjectType.LIGHTSOURCE) {
-      final String mapObjectColor = mapObject.getStringValue(MapObjectProperty.LIGHT_COLOR);
-      if (mapObjectColor != null && !mapObjectColor.isEmpty()) {
-        Color lightColor = ColorHelper.decode(mapObjectColor);
-        borderColor = new Color(lightColor.getRed(), lightColor.getGreen(), lightColor.getBlue(), 255);
-      }
-    } else if (type == MapObjectType.STATICSHADOW) {
-      borderColor = Style.COLOR_SHADOW_BORDER;
-    } else if (type == MapObjectType.SPAWNPOINT) {
-      borderColor = Style.COLOR_SPAWNPOINT;
-    } else {
-      borderColor = new Color(colorBoundingBoxFill.getRed(), colorBoundingBoxFill.getGreen(), colorBoundingBoxFill.getBlue(), 150);
-    }
-
-    g.setColor(borderColor);
-
-    RenderEngine.renderOutline(g, mapObject.getBoundingBox(), shapeStroke);
-
-    if (Program.preferences().isRenderNames()) {
-      renderName(g, mapObject);
-    }
-  }
-
-  private static void renderCollisionBox(Graphics2D g, IMapObject mapObject, BasicStroke shapeStroke) {
-    // render collision boxes
-    boolean collision = mapObject.getBoolValue(MapObjectProperty.COLLISION, false);
-    float collisionBoxWidth = mapObject.getFloatValue(MapObjectProperty.COLLISIONBOX_WIDTH, -1);
-    float collisionBoxHeight = mapObject.getFloatValue(MapObjectProperty.COLLISIONBOX_HEIGHT, -1);
-    final Align align = Align.get(mapObject.getStringValue(MapObjectProperty.COLLISION_ALIGN));
-    final Valign valign = Valign.get(mapObject.getStringValue(MapObjectProperty.COLLISION_VALIGN));
-
-    if (MapObjectType.get(mapObject.getType()) == MapObjectType.COLLISIONBOX) {
-      collisionBoxWidth = mapObject.getWidth();
-      collisionBoxHeight = mapObject.getHeight();
-      collision = true;
-    }
-
-    if (collisionBoxWidth != -1 && collisionBoxHeight != -1) {
-
-      g.setColor(Style.COLOR_COLLISION_FILL);
-      Rectangle2D collisionBox = CollisionEntity.getCollisionBox(mapObject.getLocation(), mapObject.getWidth(), mapObject.getHeight(), collisionBoxWidth, collisionBoxHeight, align, valign);
-
-      RenderEngine.renderShape(g, collisionBox);
-      g.setColor(collision ? Style.COLOR_COLLISION_BORDER : Style.COLOR_NOCOLLISION_BORDER);
-
-      Stroke collisionStroke = collision ? shapeStroke : new BasicStroke(1 / Game.world().camera().getRenderScale(), BasicStroke.CAP_ROUND, BasicStroke.JOIN_BEVEL, 0, new float[] { 1f }, 0);
-      RenderEngine.renderOutline(g, collisionBox, collisionStroke);
-    }
   }
 }
