+++ utiliti/src/de/gurkenlabs/utiliti/renderers/IEditorRenderer.java
@@ -1,0 +1,9 @@
+package de.gurkenlabs.utiliti.renderers;
+
+import java.awt.Graphics2D;
+
+public interface IEditorRenderer {
+  String getName();
+
+  void render(Graphics2D g);
+}
