+++ utiliti/src/de/gurkenlabs/utiliti/renderers/NewObjectAreaRenderer.java
@@ -1,0 +1,42 @@
+package de.gurkenlabs.utiliti.renderers;
+
+import java.awt.Font;
+import java.awt.FontMetrics;
+import java.awt.Graphics2D;
+import java.awt.geom.Rectangle2D;
+
+import de.gurkenlabs.litiengine.graphics.RenderEngine;
+import de.gurkenlabs.utiliti.Style;
+import de.gurkenlabs.utiliti.components.EditorScreen;
+import de.gurkenlabs.utiliti.components.MapComponent;
+
+public class NewObjectAreaRenderer extends MouseSelectAreaRenderer {
+
+  public NewObjectAreaRenderer() {
+    super(Style.COLOR_NEWOBJECT_FILL, Style.COLOR_NEWOBJECT_BORDER);
+  }
+
+  @Override
+  public String getName() {
+    return "NEW_OBJECT_AREA";
+  }
+
+  @Override
+  public void render(Graphics2D g) {
+    final Rectangle2D rect = EditorScreen.instance().getMapComponent().getMouseSelectArea(true);
+    if (rect == null || EditorScreen.instance().getMapComponent().getEditMode() != MapComponent.EDITMODE_CREATE) {
+      return;
+    }
+
+    this.renderSelectArea(g, rect);
+
+    g.setFont(g.getFont().deriveFont(Font.BOLD));
+
+    FontMetrics fm = g.getFontMetrics();
+
+    String width = rect.getWidth() + "";
+    String height = rect.getHeight() + "";
+    RenderEngine.renderText(g, width, rect.getX() + rect.getWidth() / 2.0 - fm.stringWidth(width) / 2.0, rect.getY() - 5);
+    RenderEngine.renderText(g, height, rect.getX() - (fm.stringWidth(height) + 3), rect.getY() + rect.getHeight() / 2);
+  }
+}
