+++ utiliti/src/de/gurkenlabs/utiliti/renderers/Renderers.java
@@ -1,0 +1,25 @@
+package de.gurkenlabs.utiliti.renderers;
+
+import java.awt.Graphics2D;
+import java.util.List;
+import java.util.concurrent.CopyOnWriteArrayList;
+
+public class Renderers {
+  private static final List<IEditorRenderer> editorRenderers;
+
+  static {
+    editorRenderers = new CopyOnWriteArrayList<>();
+    editorRenderers.add(new GridRenderer());
+    editorRenderers.add(new MapObjectsRenderer());
+    editorRenderers.add(new SelectionRenderer());
+    editorRenderers.add(new FocusRenderer());
+    editorRenderers.add(new NewObjectAreaRenderer());
+    editorRenderers.add(new MouseSelectAreaRenderer());
+  }
+
+  public static void render(Graphics2D g) {
+    for (IEditorRenderer renderer : editorRenderers) {
+      renderer.render(g);
+    }
+  }
+}
